FIX: foute references naar lessen en exercises herschreven
